Allow to use num_workers > 1 with async writer.
This is useful if many metrics have to be reported at each time step, and each of them takes a lot of time to be written (e.g. writing to disk).

It is the user's responsibility to ensure that the wrapped MetricWriter is thread-safe and can cope with out-of-order write events if they should happen.

PiperOrigin-RevId: 337396031