Makes `mask=` argument mandatory.
The problem with the current implementation is that it is too easy to forget the `mask=` argument and erroneously compute metrics with incorrect masking.

It's preferable to have metrics always require an explicit `mask=None` in the cases where no masking is applied.

Note that it is still possible to create metrics without the `mask=` argument, but  pre-defined metrics like `metrics.Average` now highlight the more safe practice of always requiring the `mask=` argument.

This change should be safe because it will only result in existing code to fail, if it does not provide the `mask=` argument.

PiperOrigin-RevId: 408651163