Adds `clu.metrics.Metric.empty()`
This allows for a more efficient pattern that has running metrics as an input/output to a jitted/pmapped eval step. In some instances (e.g. model parallel) this can substantially improve performance.

See `clu.metrics` module pydoc and `clu_synopsis.ipynb` for an illustration of this updated patttern.

Note that this change is backwards compatible. If the new pattern is to be used, then all metrics must implement the new `.empty()` classmethod.

PiperOrigin-RevId: 437040969