Improves clu.asynclib stacktraces.
The default metric writer runs asynchronously, and error messages are only surfaced the next time the metric writer is called. This leads to confusing stacktraces, and users might not understand immediately how to make sense of them.

The new error message clearly marks the stacktrace to occur asynchronously, and suggests to re-run the program with `clu.metric_writers.create_default_writer(asynchronous=False)`

PiperOrigin-RevId: 476023436