clu.metrics: Fix type annotations for `Metric`
Similar to the previous change for `Collection`, this allows the
methods and classmethods on `Metric` and its subclasses to be properly
typed. This allows us to remove the `disable=g-bare-generic` lint.

This also enables typechecking for `metrics_test.py`, which failed to
typecheck before this change.

To get `Metrics.from_fun()` to properly typecheck, we introduce a `FromFunCallable`
[Callback Protocol](https://mypy.readthedocs.io/en/stable/protocols.html#callback-protocols)
to encode the type of "a function that takes in any number of `ArrayLike`
types and outputs `Array` or a dictionary with `Array` keys".

Finally, this change uses `Protocol` instead of `abc.ABC` for `Value`,
so that we can call `v.value` on abstract `Value` types.

PiperOrigin-RevId: 529495726