Fix TPU time measurements in `PeriodicAction.timed()`.
The current code uses the time difference between the completions of two `_squareit()` calls to measure the TPU time spent in yield. But the two calls to `_squareit()` are started on a worker thread and are not synchronized with the calling thread.

```
with report.timed('foobar'):
  tpu_fn()  # <- this can slip in before start_measurement()
tpu_fn()  # <- this can slip in before stop_measurement()
```

To guarantee that the TPU programs are dispatched in the right order we need to dispatch them on the main thread. `block_until_ready()` is still called on the WorkerThread so the main thread isn't blocked.

Drive-by changes:
 - Split the two code paths (`wait_jax_async_dispatch`) as they share even less common code now.
 - No longer launder variables through `_time_per_thread`. Instead pass the start time in a future.

PiperOrigin-RevId: 535206819