Adds `preprocess_sepcs.PreprocessOps`.
The new `PreprocessOps` is a simpler version of `PreprocessFn` that allows to
have a chain of `PreprocessOp`, without mixing this behaviour with filtering
JAX types and logging transformed features. Instead, the filtering can be achieved
by simply adding the `OnlyJaxTypes`, and the logging is implemented by the new
`LogFeatures`.

`parse()` finally has been updated to use the new `PreprocessOps`, but maintain
the old behaviour (modulo string representation), with the added flag `log_features`
that allows to disable the logging functionality.

PiperOrigin-RevId: 550570417