Explicitly handle None in model parameters
As of jax 0.4.21, `jnp.array(None)` raises a warning, and in the future will result in a TypeError (see https://jax.readthedocs.io/en/latest/changelog.html#jax-0-4-21). This changes makes the current behavior explicit (i.e. `None` is converted to `NaN`) so that this deprecation in JAX will not affect clu users.

PiperOrigin-RevId: 584033892